backend/app/requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
contextlib2==21.6.0
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
hashids==1.3.1
idna==3.3
psycopg2==2.9.1
pydantic==1.8.2
pyshorteners==1.0.1
python-dotenv==0.19.1
requests==2.26.0
sniffio==1.2.0
SQLAlchemy==1.4.26
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
